Implemented working test functions for Strummer-Slider and Strummer-Picker
Refer to executeSetPickerTest and executeSetSliderTest.

executeSetPickerTest takes in a character U or D which controls which position to go to. U is upward, D is downward. To prepare for a downstrum for example, you would put the pick in the 'upward' position

executeStrumTest takes in a character, U or D, and an integer. The character determines which type of strum, Up or Down and the integer determines the number of points used for the strum. The lower the number of points, the louder the sound. We may consider setting a default for demo purposes.